The first upload
This is a first person character PC game developed using Unity 3D with
full mouse controls to look around and keyboard controls to move around.
In this game player has to shoot certain boxes that will pop up randomly
around the player. Player is always on a clock and shooting the correct
boxes will buy him time but shooting the wrong boxes will result in a
lot of time.